requirements.txt
grpcio==1.39.0
grpcio-tools==1.39.0
ipfshttpclient==0.7.0
boto3==1.18.0
cryptography==3.4.7
pyjwt==2.1.0
